feat: add dynamic form catalog books
